Updated test to read 8 bytes, rather than simply one.

+++ src/proc/syscall.c
@@ -85,14 +85,13 @@ void syscall_handle(context_t *user_context)
 void handle_syscall_read(context_t *user_context) {
     uint32_t fhandle = user_context->cpu_regs[MIPS_REGISTER_A1];
     uint8_t *buffer  = (uint8_t *) user_context->cpu_regs[MIPS_REGISTER_A2];
-    uint32_t length  = user_context->cpu_regs[MIPS_REGISTER_A3];
+    int length  = user_context->cpu_regs[MIPS_REGISTER_A3];
 
     if (fhandle == FILEHANDLE_STDIN) {
         device_t *dev = device_get(YAMS_TYPECODE_TTY, 0);
         gcd_t *gcd = (gcd_t *) dev->generic_device;
-        int len = gcd->read(gcd, buffer, (int) length);
-        // null terminate, TODO: avoid overflow
-        //buffer[len] = '\0';
+        int len = gcd->read(gcd, buffer, length);
+        // buffer[len] = '\0';
         user_context->cpu_regs[MIPS_REGISTER_V0] = len;
     } else {
         // error; reading from other files unimplemented
